requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
pymysql==1.1.0
Flask-Migrate==4.0.5
alembic==1.11.1